requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ultralytics==8.0.200
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6
python-dotenv==1.0.0
mistralai==0.1.2
httpx==0.25.0
torch==2.1.0 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cpu